requirements.txt
colorama==0.4.6
ffmpeg-python==0.2.0
future==1.0.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pytesseract==0.3.10
tqdm==4.66.2
